// ui/src/types/bindings/feed.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Feed = { uuid: string, id: string, url_feed: string, url_site: string | null, title: string | null, title_display: string | null, favicon_b64: Array<number> | null, published_at: string | null, updated_at: string | null, };